Refactor: #73 - map 페이지 하단부분 컴포넌트 분리
